Split MongoDB Workflow into its own package.
- rearranged some type classes to help with resolution

Otherwise, mostly updating imports.
