+++ mongodb/src/main/scala/quasar/physical/mongodb/MongoDbIOWorkflowExecutor.scala
@@ -21,8 +21,10 @@ import quasar.SKI._
 import quasar.{EnvironmentError, EnvErrT}
 import quasar.fs._
 import quasar.physical.mongodb.execution._
-import quasar.physical.mongodb.workflowtask._
 import quasar.physical.mongodb.mongoiterable._
+import quasar.physical.mongodb.workflow.$SortF
+import quasar.physical.mongodb.workflowtask._
+import MapReduce._
 
 import java.lang.{Boolean => JBoolean}
 import scala.Predef.classOf
@@ -36,9 +38,6 @@ import scalaz._, Scalaz._
 /** Implementation class for a WorkflowExecutor in the `MongoDbIO` monad. */
 private[mongodb] final class MongoDbIOWorkflowExecutor
   extends WorkflowExecutor[MongoDbIO, BsonCursor] {
-
-  import MapReduce._
-  import Workflow.$SortF
 
   private def foldS[F[_]: Foldable, S, A](fa: F[A])(f: (A, S) => S): State[S, Unit] =
     fa.traverseS_[S, Unit](a => MonadState[State[S,?], S].modify(f(a, _)))
