+++ mongodb/src/main/scala/quasar/physical/mongodb/WorkflowExecutor.scala
@@ -21,8 +21,10 @@ import quasar.{EnvironmentError, EnvErrT, SeqNameGeneratorT}
 import quasar.{NameGenerator => QNameGenerator, _}
 import quasar.fs.DataCursor
 import quasar.javascript._
-import quasar.physical.mongodb.workflowtask._
 import quasar.physical.mongodb.execution._
+import quasar.physical.mongodb.workflow._
+import quasar.physical.mongodb.workflowtask._
+import WorkflowExecutionError._
 
 import matryoshka._, Recursive.ops._
 import scalaz._, Scalaz._
@@ -85,9 +87,6 @@ private[mongodb] abstract class WorkflowExecutor[F[_]: Monad, C] {
   type G[A]   = ReaderT[G0, String, A]
   type M[A]   = WorkflowExecErrT[G, A]
 
-  import WorkflowExecutionError._
-  import Workflow._
-
   /** Returns a stream of the results generated by evaluating the given
     * workflow.
     *
@@ -100,7 +99,7 @@ private[mongodb] abstract class WorkflowExecutor[F[_]: Monad, C] {
     workflow: Crystallized[WorkflowF],
     defaultDb: Option[String]
   ): M[List[Bson] \/ WorkflowCursor[C]] =
-    Workflow.task(workflow) match {
+    task(workflow) match {
       case PureTask(Bson.Arr(bsons)) =>
         bsons.left[WorkflowCursor[C]].point[M]
 
@@ -131,7 +130,7 @@ private[mongodb] abstract class WorkflowExecutor[F[_]: Monad, C] {
     * (crystallized) `Workflow`.
     */
   def execute(workflow: Crystallized[WorkflowF], dst: Collection): M[Collection] =
-    execute0(Workflow task workflow, dst).run(Set()) flatMap { case (tmps, coll) =>
+    execute0(task(workflow), dst).run(Set()) flatMap { case (tmps, coll) =>
       (tmps - coll) traverse_ (c => asM(drop(c))) as coll
     }
 
@@ -328,7 +327,7 @@ private[mongodb] abstract class WorkflowExecutor[F[_]: Monad, C] {
 }
 
 object WorkflowExecutor {
-  import Workflow.{Crystallized, WorkflowF}
+  import quasar.physical.mongodb.workflow.{Crystallized, WorkflowF}
 
   /** A cursor to the result of evaluating a `Workflow`, optionally paired with
     * the cursor's temporary source that should be dropped once the cursor is
