+++ mongodb/src/main/scala/quasar/physical/mongodb/workflow/Coalesce.scala
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2014–2016 SlamData Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package quasar.physical.mongodb.workflow
+
+import quasar.Predef._
+
+import matryoshka._
+import simulacrum.typeclass
+
+@typeclass trait Coalesce[F[_]] {
+  def coalesceƒ: F[Fix[F]] => Option[F[Fix[F]]]
+
+  def coalesce: F[Fix[F]] => F[Fix[F]] = repeatedly(coalesceƒ)
+}
